refactor: Update catalog.json
This commit refactors the catalog.json file by adding a new section for SNMP and updating the formatting. The changes include adding a new "snmp" section with a list of files and dependencies. Additionally, the commit fixes a minor formatting issue by adding a newline at the end of the file.