feat: Update CONTRIBUTING.md and CONTRIBUTING.ru-RU.md
Update the CONTRIBUTING.md file to include a badge for the Russian language and add a link to the Russian version of the file. Also, create a new CONTRIBUTING.ru-RU.md file with guidelines for adding new entries in Russian.

refactor: Update README.md and README.ru-RU.md

Update the README.md file to fix a broken link to the CONTRIBUTING.md file. Also, create a new README.ru-RU.md file with information about the repository in Russian.

refactor: Update template/README.md and template/README.ru-RU.md

Update the template/README.md file to include a badge for the Russian language and add a link to the Russian version of the file. Also, create a new template/README.ru-RU.md file with instructions on how to use the template in Russian.